requirements.txt
numpy
requests
pyyaml
scapy
paramiko
pytest
flask
flask-restful
watchdog